+++ src/config/index.ts
@@ -1,2 +1,3 @@
+export * from './connection';
 export * from './escalated.config';
 export * from './user-id-column';